// web/src/api/types.ts
// THIS FILE CONTAINS APPLICATION ONLY PARAMS

// ------- Base Types -----
export type Collection = {
  id: string;
  name: string;
};

export type Pagination = {
  pageSize: number;
  page: number;
};

// ------- Responses -----

export type GetResponse<T> = {
  data: T;
};

export type ListResponse<T> = {
  data: T[];
  pagination: Pagination;
};

// ------- Requests -----

type AutoGeneratedFields = "id";

export type CreateCollectionRequest = Omit<Collection, AutoGeneratedFields>;

export type UpdateCollectionRequest = Omit<Collection, AutoGeneratedFields>;